Add leiningen project template files
Add the default files for clojure development with stub/placeholder content, as
well as a .gitignore suited for files automatically generated by leiningen.

Note that the license, which is automatically added by leiningen, has been
removed as a standalone file and from project.clj. A license will be added
later, but it may not be the EPL, as added by leiningen.